Removed redundant RFQ order tests
